Random interrupts triggering
Three different ISRs were tried to be implemented in response to external asynchronous events. However, the interrupts trigger each other for reasons unknown to me, and after many attempts, I'm forced to try a new methodology so as not to waste any more time. The next commit will be with a new solution.