2 interrupts only
A source file with only 2 interrupts was created to check if the CO2 intensity alarm wouldn't be triggered. Unfortunately after sensing the value change on the soil sensor pin and turning on the water pump the buzzer and diode are still triggered. To get rid of it I tried to use voltage stabilization, a filtering capacitor, a Schottky diode and even a different board and power supply for only powering the pump. The last option helped a lot and CO2 alarm triggering wasn't happening so many times however the problem wasn't completely resolved. After those tries it turned out that the CO2 sensor change interrupt is triggered differently every time. It should be triggered after changing the pin from high to low as well as from low to high but sometimes it changes two times as should be but the other time it is triggered only once.